% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{SpatRaster objects}
\format{
SpatRaster creating from random array. 10 cols, 10 rows, and 10 layers
}
\source{
Random values
}
\usage{
data
}
\description{
Random dataset containing binary maps of 10 species for reference and future scenarios
}
\keyword{datasets}
