% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialgn.R
\name{spatialgn}
\alias{spatialgn}
\title{Calculate average trait generalization for spatial data}
\usage{
spatialgn(x, y, trait)
}
\arguments{
\item{x}{Object of class SpatRaster with binarized distribution projected to all species in the reference scenario}

\item{y}{Object of class data.frame with species traits as columns}

\item{trait}{Column name to be used}
}
\value{
Object of class SpatRaster with average trait chosen
}
\description{
Calculate average trait generalization for spatial data
}
\examples{
set.seed(100)
mass <- runif(10, .5, 2)
elev <- sample(500:1000, 10)
ref <- terra::rast(array(sample(c(rep(1, 750), rep(0, 250))), dim = c(10, 10, 10)))
traits <- as.data.frame(cbind(mass, elev))
rownames(traits) <- paste0("sp", 1:nrow(traits))
spatialgn(ref, traits, "mass")
spatialgn(ref, traits, "elev")

}
