% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{SpatRaster and data.frame objects}
\format{
SpatRaster: 10 columns, 10 rows, and 10 layers. Data.frame: 2 columns and 10 rows.
}
\source{
Random values
}
\usage{
data
}
\description{
\emph{ref.tif} is an object of class SpatRaster created from a random array representing binary maps of 10 species in the reference scenario.
\emph{fut.tif} is an object of class SpatRaster created from a random array representing binary maps of 10 species in the future scenario.
\emph{traits.xlsx} is an object of class data.frame created from random vectors representing species traits: body mass and elevation.
}
\keyword{datasets}
