% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdv.R
\name{spatialdv}
\alias{spatialdv}
\title{Calculate diversity patterns for spatial data}
\usage{
spatialdv(x, y, z = "delta")
}
\arguments{
\item{x}{Object of class SpatRaster with binarized distribution projected to all species in the reference scenario}

\item{y}{Object of class SpatRaster with binarized distribution projected to all species in the future scenario}

\item{z}{Index to be return}
}
\value{
Plot object of class SpatRaster according to the index chosen
}
\description{
Calculate alpha and beta diversity for spatial data
}
\examples{
require(terra)
set.seed(100)
ref <- rast(array(sample(c(rep(1, 750), rep(0, 250))), dim = c(10, 10, 10)))
fut <- rast(array(sample(c(rep(1, 300), rep(0, 700))), dim = c(10, 10, 10)))
spatialdv(ref, fut, "beta")

}
